# apps/fox-deck-api/src/modules/resource/dto/get-resource-root-by-user-id.dto.ts
import {ApiProperty} from "@nestjs/swagger";
import {IsString} from "class-validator";


export class GetResourceRootByUserIdResponseDto {
    @ApiProperty()
    @IsString()
    resourceId: string

    @ApiProperty()
    @IsString()
    parentResourceId: string

    @ApiProperty()
    @IsString()
    type: string

    @ApiProperty()
    @IsString()
    name: string

    @ApiProperty()
    @IsString()
    content: string

    @ApiProperty()
    @IsString()
    createdAt: string
}