# apps/fox-deck-api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/login":{"post":{"operationId":"UserController_getUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}}},"responses":{"200":{"description":"A JWT Token containing basic user information.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"tags":["Users"]}},"/register":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/create-resource":{"post":{"operationId":"ResourceController_createResource","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateResourceRequestDto"}}}},"responses":{"200":{"description":"The created resource resourceId.","schema":{"$ref":"#/components/schemas/CreateResourceRequestDto"},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateResourceRequestDto"}}}}},"tags":["Resources"],"security":[{"access-token":[]}]}},"/resource":{"get":{"operationId":"ResourceController_getResourceByUserId","parameters":[],"responses":{"200":{"description":"The resources of the logged in user.","schema":{"$ref":"#/components/schemas/GetResourceByUserIdResponseDto"},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetResourceByUserIdResponseDto"}}}}}},"tags":["Resources"],"security":[{"access-token":[]}]}},"/resource-root":{"get":{"operationId":"ResourceController_getRootLevelResourceByUserId","parameters":[],"responses":{"200":{"description":"The root-level resources of the logged in user.","schema":{"$ref":"#/components/schemas/GetResourceRootByUserIdResponseDto"},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetResourceRootByUserIdResponseDto"}}}}}},"tags":["Resources"],"security":[{"access-token":[]}]}}},"info":{"title":"FoxDeck API","description":"This API handles requests from the FoxDeck Applications","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"LoginRequestDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateUserRequestDto":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}},"required":["email","username","password"]},"CreateResourceRequestDto":{"type":"object","properties":{"parentResourceId":{"type":"string"},"name":{"type":"string"},"content":{"type":"string"},"type":{"type":"string","enum":["note","folder"]}},"required":["parentResourceId","name","content","type"]},"GetResourceByUserIdResponseDto":{"type":"object","properties":{"resourceId":{"type":"string"},"parentResourceId":{"type":"string"},"type":{"type":"string"},"name":{"type":"string"},"content":{"type":"string"},"createdAt":{"type":"string"}},"required":["resourceId","parentResourceId","type","name","content","createdAt"]},"GetResourceRootByUserIdResponseDto":{"type":"object","properties":{"resourceId":{"type":"string"},"parentResourceId":{"type":"string"},"type":{"type":"string"},"name":{"type":"string"},"content":{"type":"string"},"createdAt":{"type":"string"}},"required":["resourceId","parentResourceId","type","name","content","createdAt"]}}}}